Include some of FastUtil. expand, fringe added
Grazie, Sebastiano Vigna! I hope that including the source from FastUtil
isn't in any way a breach of the Apache license that both of these
projects use. I was concerned about the 16MB jar size for FastUtil as a
whole, which is why just the part that I think I might use is included
here.